dom/workers/test/test_chromeWorker.xul

<?xml version="1.0"?>
<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<window title="DOM Worker Threads Test"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="test();">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script type="application/javascript" src="dom_worker_helper.js"/>

  <script type="application/javascript">
  <![CDATA[

    function test()
    {
      waitForWorkerFinish();

      var worker = new ChromeWorker("chromeWorker_worker.js");
      worker.onmessage = function(event) {
        is(event.data, "Done!", "Wrong message!");
        finish();
      }
      worker.onerror = function(event) {
        ok(false, "Worker had an error: " + event.message);
        worker.terminate();
        finish();
      }
      worker.postMessage("go");
    }

  ]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display:none;"></div>
    <pre id="test"></pre>
  </body>
  <label id="test-result"/>
</window>
